Update README.md
Tic-tac-toe is a straightforward two-player game that, if both players play their best, will always end in a tie. The game is also known as Xs and Os or zeros and crosses.

A computer or other device can be used to play the game of tic tac toe, which is typically played by drawing on paper. This timeless game provides the foundation for other ones, like Connect 4.

This game provides the following features:

This game is developed on a 3×3 grid.
This game is designed for two players.
Every player may choose a letter between X and O.
Both players will get their chances to turn by turn.